+++ packages/react-hooks/src/index.ts
@@ -2,4 +2,4 @@ export * from './MedplumProvider/MedplumProvider';
 export * from './MedplumProvider/MedplumProvider.context';
 export * from './useResource/useResource';
 export * from './useSearch/useSearch';
-export * from './useCachedS3BinaryUrl/useCachedS3BinaryUrl';
+export * from './useCachedBinaryUrl/useCachedBinaryUrl';
