+++ packages/react/src/AttachmentDisplay/AttachmentDisplay.tsx
@@ -1,6 +1,6 @@
 import { Anchor } from '@mantine/core';
 import { Attachment } from '@medplum/fhirtypes';
-import { useCachedS3BinaryUrl } from '@medplum/react-hooks';
+import { useCachedBinaryUrl } from '@medplum/react-hooks';
 
 export interface AttachmentDisplayProps {
   value?: Attachment;
@@ -9,7 +9,7 @@ export interface AttachmentDisplayProps {
 
 export function AttachmentDisplay(props: AttachmentDisplayProps): JSX.Element | null {
   const { contentType, url: uncachedUrl, title } = props.value ?? {};
-  const url = useCachedS3BinaryUrl(uncachedUrl);
+  const url = useCachedBinaryUrl(uncachedUrl);
 
   if (!url) {
     return null;
