+++ packages/react/src/ResourceAvatar/ResourceAvatar.tsx
@@ -1,7 +1,7 @@
 import { Avatar, AvatarProps } from '@mantine/core';
 import { getDisplayString, getImageSrc } from '@medplum/core';
 import { Reference, Resource } from '@medplum/fhirtypes';
-import { useCachedS3BinaryUrl, useResource } from '@medplum/react-hooks';
+import { useCachedBinaryUrl, useResource } from '@medplum/react-hooks';
 import { MedplumLink } from '../MedplumLink/MedplumLink';
 
 export interface ResourceAvatarProps extends AvatarProps {
@@ -13,7 +13,7 @@ export function ResourceAvatar(props: ResourceAvatarProps): JSX.Element {
   const resource = useResource(props.value);
   const text = resource ? getDisplayString(resource) : props.alt ?? '';
   const uncachedImageUrl = (resource && getImageSrc(resource)) ?? props.src;
-  const imageUrl = useCachedS3BinaryUrl(uncachedImageUrl ?? undefined);
+  const imageUrl = useCachedBinaryUrl(uncachedImageUrl ?? undefined);
   const radius = props.radius ?? 'xl';
 
   const avatarProps = { ...props };
